moved getAllEvents to AccountManager instead of EventManager and added getEvent to EventManager
